Update2.0 hangman.asm
I finally managed to implement a tracker for the number of attempts remaining in the code. Also, I managed to successfully print out all found chars within the string while printing the blank spaces. Yet, I'm still hardcoding the guessing word since I'm not 100% sure how to get around that. So that's one area that needs to be addressed.