Create updatedHangman.asm
One Last Hurrah!!! Hey guys, this is an updated version of our final project. I pretty much worked on this after the date we submitted the original version. You can scroll to the bottom to see the things I managed to fix and implement newly. I managed to actually draw hangmen, I made it so that the user gets a new word after they retry, without having to reassemble the code for a newly-generated word, and tons of more updates. This is more of a personal-gain sort of thing, but I'll leave it here for you. This was awesome boys, nice work on our project!